refactor: move Codex parser into adapter
